+++ mission/views.py
@@ -20,8 +20,6 @@ from assets.models import Asset, AssetStatus
 from mission.helpers import get_my_assets_not_in_mission
 from organization.decorators import asset_is_operator, get_organization_from_id
 from organization.models import Organization, OrganizationMember, OrganizationAsset
-from search.models import Search
-from search.view_helpers import check_searches_in_progress
 from timeline.models import TimeLineEntry
 from timeline.helpers import timeline_record_create, \
     timeline_record_external_reference_add, timeline_record_external_reference_remove, timeline_record_external_reference_update, \
@@ -31,7 +29,7 @@ from timeline.helpers import timeline_record_create, \
 
 from .models import Mission, MissionExternalReference, MissionUser, MissionAsset, MissionAssetType, MissionOrganization, MissionAssetStatus, MissionAssetStatusValue, AssetCommand
 from .forms import AssetCommandForm, MissionForm, MissionUserForm, MissionAssetForm, MissionOrganizationForm
-from .decorators import get_user_from_id, mission_asset_get, mission_can_add_organization, mission_can_add_user, mission_is_member, mission_is_admin
+from .decorators import get_user_from_id, mission_can_add_organization, mission_can_add_user, mission_is_member, mission_is_admin
 
 
 @method_decorator(login_required, name="dispatch")
@@ -714,34 +712,6 @@ class MissionExternalReferenceView(View):
         extref.save()
         timeline_record_external_reference_remove(mission=mission_user.mission, user=mission_user.user, external_reference=extref)
         return HttpResponse("Done")
-
-
-@method_decorator(login_required, name="dispatch")
-@method_decorator(asset_is_operator, name="dispatch")
-class AssetMissionView(View):
-    """
-    Mission/search context for an asset: last command, active mission, and searches.
-    """
-    def get(self, request, asset):
-        """
-        Return mission/search context for this asset as JSON.
-        """
-        data = {
-            'last_command': AssetCommand.last_command_for_asset_to_json(asset),
-        }
-        mission_asset = mission_asset_get(asset)
-        if mission_asset is not None:
-            data['mission_id'] = mission_asset.mission.pk
-            data['mission_name'] = mission_asset.mission.mission_name
-
-            current_search = check_searches_in_progress(mission_asset.mission, asset)
-            if current_search is not None:
-                data['current_search_id'] = current_search.pk
-            queued_search = Search.oldest_queued_for_asset(mission_asset.mission, asset)
-            if queued_search is not None:
-                data['queued_search_id'] = queued_search.pk
-
-        return JsonResponse(data)
 
 
 @method_decorator(login_required, name='dispatch')
